pyproject.toml
[project]
name = "inbox-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1",
    "click>=8.1,<9",
    "rich>=13.9,<14",
    "python-dotenv>=1.0,<2",
]
license = "AGPL-3.0-or-later"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
inbox-agent = "inbox_agent.cli:cli"
